Decouple loc files. Remove commented code
