+++ src/CarbonAware.Aggregators/test/TestData.cs
@@ -86,8 +86,7 @@ public static class TestData
         return new EmissionsForecast(){
             GeneratedAt = DateTimeOffset.Parse("2022-01-01T00:00:00Z"),
             Location = new Location() {
-                RegionName = "westus",
-                // CloudProvider = CloudProvider.Azure
+                RegionName = "westus"
             },
             ForecastData = forecastData
         };
