+++ src/CarbonAware.LocationSources/test/LocationSourceTest.cs
@@ -19,7 +19,6 @@ public class LocationSourceTest
         var mockLocationSource = SetupMockLocationSource().Object;
         Location inputLocation = new Location {
             LocationType = LocationType.CloudProvider,
-            // CloudProvider = CloudProvider.Azure,
             RegionName = "eastus"
         };
 
@@ -28,7 +27,6 @@ public class LocationSourceTest
 
         inputLocation = new Location {
             LocationType = LocationType.CloudProvider,
-            // CloudProvider = CloudProvider.Azure,
             RegionName = "westus"
         };
